+++ build.sh
@@ -1,4 +1,4 @@
-TAG="1.2.0.beta9"
+TAG="1.3.0.alpha5"
 
 docker build . -t  andonisalcedo/podinsights:$TAG
 docker push andonisalcedo/podinsights:$TAG
